inject asset details in asset controller
